# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shift API
  description: |
    One API serving both the Shift web app and the future React Native app,
    with zero platform-specific branching in the API layer.

    * Auth uses Bearer access tokens (short-lived JWT) + rotating refresh tokens.
    * All mutating `POST`/`PATCH` requests accept an `Idempotency-Key` header.
    * All timestamps are ISO 8601 strings in UTC.
    * All IDs are UUIDv4 strings.
    * List endpoints that can grow unbounded use cursor-based pagination.
  version: "1.0.0"
  contact:
    name: Shift
servers:
  - url: https://api.shift.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
                name: { type: string }
      responses:
        "200":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/Error" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email + password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh token for a new access/refresh pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New token pair issued; old refresh token invalidated
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  refreshToken: { type: string }
        "401": { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out (revoke a single refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "204": { description: Logged out }
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "204":
          description: Always 204, regardless of whether the email exists
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using a reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, format: password, minLength: 8 }
      responses:
        "204": { description: Password updated }
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify an email address using a verification token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "204": { description: Email verified }
  /auth/oauth/{provider}/start:
    get:
      tags: [Auth]
      summary: Begin an OAuth flow (sign-in or connect-account)
      security: []
      parameters:
        - $ref: "#/components/parameters/OAuthProvider"
      responses:
        "302": { description: Redirect to the OAuth provider }
  /auth/oauth/{provider}/callback:
    get:
      tags: [Auth]
      summary: OAuth provider callback
      security: []
      parameters:
        - $ref: "#/components/parameters/OAuthProvider"
      responses:
        "200":
          description: |
            Web: redirected with tokens in a one-time-use URL fragment.
            RN: deep-linked back into the app (e.g. `shift://oauth-callback?code=...`),
            exchanged for tokens in a follow-up POST.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }

  # ---------------------------------------------------------------- GOALS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived (15 min) access token. Refresh via POST /auth/refresh.
  parameters:
...
    OAuthProvider:
      name: provider
      in: path
      required: true
      schema: { type: string, enum: [github, google, figma] }
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: TASK_NOT_FOUND }
            message: { type: string, example: "No task with that id on this goal's roadmap." }
            requestId: { type: string, example: "req_8f2a..." }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        emailVerifiedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        accessToken: { type: string }
        refreshToken: { type: string }
...